bind: support for buffer-protocol
